Move Design out of Sidebar folder
