+++ packages/story-editor/src/components/design/updateProperties.js
@@ -16,7 +16,7 @@
 /**
  * Internal dependencies
  */
-import { MULTIPLE_VALUE } from '../../../constants';
+import { MULTIPLE_VALUE } from '../../constants';
 
 /**
  * @param {Object} currentProperties The existing element properties.
