+++ packages/story-editor/src/components/panels/shared/test/_utils.js
@@ -28,7 +28,7 @@ import { useMemo } from '@googleforcreators/react';
 import theme from '../../../../theme';
 import useHandlers from '../../../../utils/useHandlers';
 import FormContext from '../../../form/context';
-import updateProperties from '../../../sidebar/design/updateProperties';
+import updateProperties from '../../../design/updateProperties';
 
 function TestPanel({
   panelType,
